chore: move server utils into api-server
